% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gram.dis}
\alias{gram.dis}
\title{gram.dis}
\usage{
gram.dis(y)
}
\arguments{
\item{y}{discrete vector}
}
\value{
gram matrix for discrete vector
}
\description{
gram.dis
}
\examples{
toy <- c(1,2,3,1)
result=gram.dis(toy)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
