% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gram.gauss}
\alias{gram.gauss}
\title{gram.gauss}
\usage{
gram.gauss(x,x.new,complexity)
}
\arguments{
\item{x}{previous observations}

\item{x.new}{new observations}

\item{complexity}{tuning parameter in Gaussian kernel}
}
\value{
Gram matrix for the Gaussian kernel. This also can be used to project predictor on the testing set.
}
\description{
gram.gauss
}
\examples{
old <- matrix(c(1,2,3,4),2,2)
new <- matrix(c(5,6,7,8),2,2)
result <- gram.gauss(old,new,1)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
