% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpca.R
\name{kpca}
\alias{kpca}
\title{kpca}
\usage{
kpca(x,complexity)
}
\arguments{
\item{x}{dataset}

\item{complexity}{tuning parameter in Gaussian kernel. larger complexity means a wiggly kernel function}
}
\value{
principal component
}
\description{
kpca
}
\examples{
n = 50; p = 5
x = matrix(rnorm(n*p),n,p)
pred=kpca(x,1)[,1:3]
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
