% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matpower.R
\name{matpower}
\alias{matpower}
\title{matpower}
\usage{
matpower(a,alpha)
}
\arguments{
\item{a}{matrix}

\item{alpha}{power}
}
\value{
power of a matrix
}
\description{
matpower
}
\examples{
mat <- matrix(rnorm(4,1,3),2,2)
invmat <- matpower(mat,-1)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
