% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mppower}
\alias{mppower}
\title{mppower}
\usage{
mppower(matrix,power,ignore)
}
\arguments{
\item{matrix}{input matrix}

\item{power}{power}

\item{ignore}{ignoring criterion}
}
\value{
Moore penrose inverse
}
\description{
mppower
}
\examples{
a <- matrix(rnorm(4,0,1),2,2)
mppower(a,-1,0.2)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
