% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{onorm}
\alias{onorm}
\title{onorm}
\usage{
onorm(a)
}
\arguments{
\item{a}{input matrix}
}
\value{
result of the operator norm
}
\description{
onorm
}
\examples{
a <- matrix(c(1,2,3,4),2,2)
result <- onorm(a)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
