% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgepower.R
\name{ridgepower}
\alias{ridgepower}
\title{ridgepower}
\usage{
ridgepower(a,e,c)
}
\arguments{
\item{a}{square matrix}

\item{e}{tuning parameter}

\item{c}{power}
}
\value{
matrix with the power
}
\description{
ridgepower
}
\examples{
x <- matrix(c(1:4),2,2)
result <- ridgepower(x, 0.001, -1)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
