% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spearman}
\alias{spearman}
\title{standmat}
\usage{
spearman(x1,x2)
}
\arguments{
\item{x1}{first argument}

\item{x2}{second argument}
}
\value{
standardized matrix
}
\description{
standmat
}
\examples{
x1 <- rnorm(100)
x2 <- rnorm(100)
spearman(x1,x2)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
