% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standmat.R
\name{standmat}
\alias{standmat}
\title{standmat}
\usage{
standmat(x)
}
\arguments{
\item{x}{matrix}
}
\value{
standardized matrix
}
\description{
standmat
}
\examples{
mat <- matrix(rnorm(4),2,2)
standmat(mat)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
