% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sym}
\alias{sym}
\title{sym}
\usage{
sym(a)
}
\arguments{
\item{a}{any matrix}
}
\value{
symmetrize matrix when matrix is theoretically symmetric but not in numerically
}
\description{
sym
}
\examples{
ex <- matrix(c(1.1,2.1,1.2,2.2),2,2)
result <- sym(ex)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
