% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{tr}
\alias{tr}
\title{trace}
\usage{
tr(a)
}
\arguments{
\item{a}{any matrix}
}
\value{
trace value of the matrix
}
\description{
trace
}
\examples{
mat <- matrix(2,2,2)
tr(mat)
}
\references{
Li, B. (2018). Sufficient dimension reduction: Methods and applications with R. CRC Press.
}
