% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Chemical ingredients of wine dataset}
\format{
A data frame with 178 observations on 14 variables.

Attributes
1. cultivar
2. alcohol
3. malic acid
4. ash
5. alcalinity of ash
6. magnesium
7. total phenols
8. flavanoids
9. nonflavanoid phenols
10. oroanthocyanins
11. color intensity
12. hue
13. OD280/OD315 of diluted wines
14. proline.
}
\source{
https://archive.ics.uci.edu/ml/datasets/wine
}
\usage{
wine
}
\description{
This dataset is about chemical analysis of 178 wines in a particular region of Italy. There are three cultivars with 59, 71, and 48, respectively.
}
\references{
Forina, M., Leardi, R., Armanino, C., Lanteri, S., Conti, P., & Princi, P. (1988). PARVUS: An extendable package of programs for data exploration, classification and correlation. Journal of Chemometrics, 4(2), 191-193..
}
\keyword{datasets}
